pyproject.toml
[tool.poetry]
name = "DisTelCon-Bot"
version = "0.1.0"
description = ""
authors = ["Dario Caricchio <dariocaricchio@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
discord = "^1.0.1"
flask = "^1.1.2"
wikiquote = "^0.1.14"
lxml = "^4.6.2"
replit = "^2.0.0"
pyTelegramBotAPI = "^3.7.6"
python-dotenv = "^0.15.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
